+++ src/generators/boolean/index.ts
@@ -1,0 +1,8 @@
+import {Chance} from 'chance';
+
+/**
+ * Returns `true` or `false`, chosen at random.
+ */
+const createBooleanGenerator = (chance: Chance.Chance) => (): boolean => chance.bool();
+
+export default createBooleanGenerator;