+++ src/generators/first_name/index.ts
@@ -1,0 +1,8 @@
+import {Chance} from 'chance';
+
+/**
+ * Returns a randomly-selected first name, e.g. `Carter`.
+ */
+const createFirstNameGenerator = (chance: Chance.Chance) => (): string => chance.first();
+
+export default createFirstNameGenerator;